Create PDF Engine
PDF Engine :

This is based on existing PDF box parser. Extending the capabilities provided by the PDF parser to add new features. 

Some times to extract key information present in the PDF document it becomes necessary to identify the attributes pertaining to each word in the PDF file for better understanding of patterns observed in the PDF file.  PDF engine will do the exactly same. It provides you they position of each words for each file present in the PDF document.